requirements.txt
pillow
numpy
matplotlib
histogram
tkinter
streamlit
os
base64
io

